perf: use `Math.random()` because it's probably faster and security doesn't matter
